+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- separate heartbeat and message modes in HTTP API
+- support separate heartbeat and message modes via HTTP API
 
 ## 0.1.9 - 2026-03-05